adding color and animation constants
