+++ Powerup/AnimationConstants.swift
@@ -1,0 +1,26 @@
+import Foundation
+import SpriteKit
+// Animations
+let swappingAnimationDuration = 0.2
+let endSceneFadeInAnimationDuration = 0.5
+let clipboardStarBlinkAnimationDuration = 0.3
+// Animation constants.
+let boxEnlargingScale = CGFloat(1.2)
+let boxEnlargingDuration = 0.25
+let buttonWaitDuration = 0.5
+let boxFlipInterval = 0.2
+let showAllBoxesInterval = 0.3
+let boxDarkening = SKAction.colorize(with: UIColor(white: 0.6, alpha: 0.8), colorBlendFactor: 1.0, duration: 0.2)
+let fadeInAction = SKAction.fadeIn(withDuration: 0.8)
+let fadeOutAction = SKAction.fadeOut(withDuration: 0.8)
+let scoreTextPopScale = CGFloat(1.2)
+let scoreTextPopDuraion = 0.25
+// Animations
+let questionFadeInTime = 0.3
+let questionFadeOutTime = 0.2
+let correctWrongSpriteStayTime = 0.5
+let boatRaiseDuration = 0.4
+let boatRotateAngle = 20.0
+let boatRotateDuration = 4.0
+let endScenePauseDuration = 1.0
+let endSceneFadeInDuration = 1.0
