+++ Powerup/Constants.swift
@@ -10,5 +10,9 @@ let descriptionTextPosX = 0.19
 let descriptionTextPosY = -0.1
 // MARK: Game Constants MinesweeperGameScene.swift
 let gridSizeCount = 5
+// How many boxes could be selected each round.
+let selectionMaxCount = 5
 // MARK: Constants SinkToSwimGameScene.swift
 let sinkingSpeedRelativeToGauge = 0.025
+// TODO: Should detemine how many Karma points will be given after each completion of scenario.
+let karmaGain = 20
