+++ Powerup/ResultsViewController.swift
@@ -1,10 +1,7 @@
 import UIKit
 
 class ResultsViewController: UIViewController {
-    
-    // TODO: Should detemine how many Karma points will be given after each completion of scenario.
-    let karmaGain = 20
-    
+        
     // MARK: Properties
     // This will be set in the ScenarioViewController.
     var completedScenarioID: Int = -1
